[hotfix][runtime] Move active resource manager related classes to separate package.
